+++ SliceyDicey.Lib/Gcode/GcodeParser.cs
@@ -18,14 +18,18 @@ public static class GcodeParser
             lineNo += 1;
             var line = await reader.ReadLineAsync().ConfigureAwait(false);
             if (string.IsNullOrWhiteSpace(line)) continue;
-            if (line is null) break;
 
             if (line.StartsWith(GcodePngThumbnail.Prefix))
             {
                 var parts = line.Split(' ');
                 thumbnail = new GcodePngThumbnail(parts[3], long.Parse(parts[4]));
             }
-            else if (thumbnail is not null && line.StartsWith(GcodeThumbnail.Suffix))
+            else if (line.StartsWith(GcodeQoiThumbnail.Prefix))
+            {
+                var parts = line.Split(' ');
+                thumbnail = new GcodeQoiThumbnail(parts[3], long.Parse(parts[4]));
+            }
+            else if (thumbnail is not null && (line.StartsWith(GcodeQoiThumbnail.Suffix) || line.StartsWith(GcodePngThumbnail.Suffix)))
             {
                 thumbnails.Add(thumbnail);
                 thumbnail = null;